backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
python-multipart 